+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]## [0.5.2.1b]
 ### Changed
+- (critical) Fix IPython import for `_show()` workaround in PlotlyManager to avoid import exception when running on local workstation
 
 ## [0.5.2.0] - 2021-12-26
 ### Changed